Configure RoomDB to fallback to destructive migration when a migratio… (#1273)
* Configure RoomDB to fallback to destructive migration when a migration is not provided.
We do not provide a migration path from a higher number to a lower number and in those cases, the DB is cleared out instead of the host app crashing.
Since the DB we use today only handles the analytics events and they are uploaded every 30 secs, the risk of dropping data should be minimal.

* Add CHANGELOG.md

* WIP -- add tests

* Add tests

* Remove unneeded test

* add a comment

* Update CHANGELOG.md

Co-authored-by: Sarah Koop <skoop@paypal.com>

---------

Co-authored-by: Sarah Koop <skoop@paypal.com>